+++ kubeconfig.tf
@@ -8,7 +8,7 @@ resource "ssh_sensitive_resource" "kubeconfig" {
   bastion_user        = local.ssh_bastion.bastion_user
   bastion_private_key = local.ssh_bastion.bastion_private_key
 
-  host        = can(ipv6(local.first_control_plane_ip)) ? "[${local.first_control_plane_ip}]" : local.first_control_plane_ip
+  host        = provider::assert::ipv6(local.first_control_plane_ip) ? "[${local.first_control_plane_ip}]" : local.first_control_plane_ip
   port        = var.ssh_port
   user        = "root"
   private_key = var.ssh_private_key
@@ -44,12 +44,12 @@ locals {
     :
     (can(local.first_control_plane_ip) ? local.first_control_plane_ip : "unknown")
   )
-  kubeconfig_server_host = can(ipv6(local.kubeconfig_server_address)) ? "[${local.kubeconfig_server_address}]" : local.kubeconfig_server_address
+  kubeconfig_server_host = provider::assert::ipv6(local.kubeconfig_server_address) ? "[${local.kubeconfig_server_address}]" : local.kubeconfig_server_address
   kubeconfig_server      = "https://${local.kubeconfig_server_host}:${var.kubeapi_port}"
   kubeconfig_external = replace(
     replace(
       ssh_sensitive_resource.kubeconfig.result,
-      "https://127.0.0.1:6443",
+      "https://127.0.0.1:${var.kubeapi_port}",
       local.kubeconfig_server
     ),
     "default",